requirements.txt
fastapi
uvicorn
python-multipart
pillow
requests
